develop last More block: seeds page, seed_info page, chemicals page, chemical_info page, fertilizers page, fertilizer_info page, land_parcels page, land_parcel_info page, user_info page, additional_objects page, fix some bugs with json objects and some other models fields

+++ src/main/java/com/agro/agro_project/dao/FertilizerRepository.java
@@ -8,8 +8,8 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository("fertilizerRepository")
-public interface FertilizerRepository extends JpaRepository<Fertilizer, Integer> {
-    Optional<Fertilizer> findById(Integer id);
+public interface FertilizerRepository extends JpaRepository<Fertilizer, Short> {
+    Optional<Fertilizer> findById(Short id);
 
     List<Fertilizer> findAll();
 }
